pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "shin rakuda"
version = "0.1.0"
description = "Japanese LLM Benchmark Tool"
authors = ["YuzuAI <contact@yuzuai.jp>"]

[tool.poetry.dependencies]
python = "^3.9"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"
pydantic = "^2.6.4"
openai = "^1.14.2"
google-generativeai = "^0.4.1"
anthropic = "^0.21.3"
transformers = "^4.40.0"
torch = "^2.1.2"
accelerate = "^0.28.0"
tqdm = "^4.66.3"
types-tqdm = "^4.66.0.20240106"
python-dotenv= "^1.0.1"
sentencepiece= "^0.2.0"
matplotlib = "^3.8.4"
tenacity= "^8.2.3"
vllm="^0.4.0.post1"
filelock="^3.14.0"

[tool.poetry.dev-dependencies]
